Move Environment, DynamicScope, and Iterable to glimmer-application.
Clone the current versions as test helpers so this repo can continue to
test rendering.
