+++ src/index.ts
@@ -3,5 +3,3 @@ export { default as ComponentFactory } from './component-factory';
 export { default as ComponentDefinition } from './component-definition';
 export { default as ComponentLayoutCompiler } from './component-layout-compiler';
 export { default as ComponentManager } from './component-manager';
-export { default as DynamicScope } from './dynamic-scope';
-export { default as Environment, EnvironmentOptions } from './environment';
