--- test/environment-test.ts
@@ -1,35 +1,0 @@
-import { getOwner, setOwner, Owner } from '@glimmer/di';
-import { DOMTreeConstruction } from '@glimmer/runtime';
-import Environment, { EnvironmentOptions } from '../src/environment';
-
-const { module, test } = QUnit;
-
-module('Environment');
-
-test('can be instantiated with new', function(assert) {
-  let env = new Environment({
-    document: self.document,
-    appendOperations: new DOMTreeConstruction(self.document)
-  });
-  assert.ok(env, 'environment exists');
-});
-
-test('can be instantiated with create', function(assert) {
-  let env = Environment.create();
-  assert.ok(env, 'environment exists');
-});
-
-test('can be assigned an owner', function(assert) {
-  class FakeApp implements Owner {
-    identify(specifier: string, referrer?: string) { return ''; }
-    factoryFor(specifier: string, referrer?: string) { return null; }
-    lookup(specifier: string, referrer?: string) { return null; }
-  }
-  let app = new FakeApp;
-
-  let options: EnvironmentOptions = {};
-  setOwner(options, app);
-
-  let env = Environment.create(options);
-  assert.strictEqual(getOwner(env), app, 'owner has been set');
-});
